split error into directory
